report: by_{month,year}() return list of dicts
Add option as_dataframe (default True for now) to allow transition to
using JSON as return values. We will be using JSON data for the
metricsgraphics output.

Also fixed the monthly grouping, previously it was not considering the
year in the month grouping which would have led to errors if we had
multiple years in our dataset.